backend/pyproject.toml
[project]
name = "fitness-api-pet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "fastapi>=0.125.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
    "PyJWT>=2.8.0",
    "sqlmodel>=0.0.24",
    "pydantic-settings>=2.6.1",
    "redis>=5.0.0"
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
]

[tool.ruff]
# Длина строки по умолчанию
line-length = 100
# Исключаемые директории
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "env",
    ".env",
    "alembic/versions",
    "*.pyc",
    "../frontend",
]

[tool.ruff.lint]
# Включаемые правила
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
# Игнорируемые правила
ignore = [
    "E501",  # line too long (управляется через line-length)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
# Использовать ruff для сортировки импортов
known-first-party = ["app"]

[tool.ruff.format]
# Использовать ruff для форматирования
quote-style = "double"
indent-style = "space"

[tool.uv.workspace]
members = [
    "fitness-api",
]
